+++ R/pkg-sf.R
@@ -126,12 +126,23 @@ wk_crs_equal_generic.crs <- function(x, y, ...) {
 wk_crs_proj_definition.crs <- function(crs, proj_version = NULL, verbose = FALSE) {
   if (is.na(crs)) {
     wk_crs_proj_definition(NULL)
-  } else if (verbose) {
+  } else if (isTRUE(verbose)) {
     crs$Wkt %||% crs$wkt
   } else if (isTRUE(is.na(crs$epsg)) || isTRUE(grepl("^[0-9A-Za-z]+:[0-9A-Za-z]+$", crs$input))) {
     wk_crs_proj_definition(crs$input)
   } else {
     paste0("EPSG:", crs$epsg)
+  }
+}
+
+#' @export
+wk_crs_projjson.crs <- function(crs) {
+  json <- crs$ProjJson
+  if (is.null(json)) {
+    # i.e., GDAL is not >= 3.1.0
+    NextMethod()
+  } else {
+    json
   }
 }