+++ tests/testthat/test-pkg-sf.R
@@ -25,6 +25,11 @@ test_that("wk_crs_proj_definition() works for sf crs objects", {
   )
 })
 
+test_that("wk_crs_projjson() works for sf crs objects", {
+  expect_match(wk_crs_projjson(sf::st_crs(4326)), "GeographicCRS")
+  expect_identical(wk_crs_projjson(sf::NA_crs_), NA_character_)
+})
+
 test_that("wk_crs/set_crs works on sf/sfc", {
   skip_if_not_installed("sf")